+++ src/main/java/cn/memoryzy/json/action/transform/ToPropertyAction.java
@@ -1,8 +1,0 @@
-package cn.memoryzy.json.action.transform;
-
-/**
- * @author Memory
- * @since 2024/11/13
- */
-public class ToPropertyAction {
-}