+++ src/main/java/cn/memoryzy/json/model/strategy/clipboard/UrlParamConversionStrategy.java
@@ -1,7 +1,7 @@
 package cn.memoryzy.json.model.strategy.clipboard;
 
 import cn.memoryzy.json.model.strategy.clipboard.context.ClipboardTextConversionStrategy;
-import cn.memoryzy.json.util.TextTransformUtil;
+import cn.memoryzy.json.util.DataConverter;
 
 /**
  * @author Memory
@@ -11,12 +11,12 @@ public class UrlParamConversionStrategy implements ClipboardTextConversionStrate
 
     @Override
     public boolean canConvert(String text) {
-        return null != TextTransformUtil.urlParamsToJson(text);
+        return null != DataConverter.urlParamsToJson(text);
     }
 
     @Override
     public String convertToJson(String text) {
-        return TextTransformUtil.urlParamsToJson(text);
+        return DataConverter.urlParamsToJson(text);
     }
 
 }