+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 ## [Unreleased]
 - 完全支持 [JSON5](https://json5.org/)。
-- 新增 URL Param 转为 JSON 文本。
+- 新增 URL 参数 与 JSON 互转。
+- 新增 Properties 格式参数与 JSON 互转。
 - 新增 JSON 转义功能。
 - 用户体验优化与改进。
 - BUG修复。
 - Full support for [JSON5](https://json5.org/).
-- Added URL Param converted to JSON text.
+- Added URL parameter interconversion with JSON.
+- Added the Properties format parameter to interconvert with JSON.
 - Added JSON escape function.
 - User experience optimization and improvement.
 - BUG fixes.